feat: refactor configuration and logging setup; enhance database initialization logic

+++ edi/02-concesionaria/api/core/db.py
@@ -1,4 +1,7 @@
-from sqlmodel import create_engine, SQLModel, Session
+import os
+from core.logger import logger
+from sqlmodel import create_engine, SQLModel, Session, inspect
+from models.car import Car
 
 sqlite_file_name = "database.db"
 sqlite_url = f"sqlite:///{sqlite_file_name}"
@@ -19,9 +22,19 @@ def get_session():
         yield session
 
 
-def create_db_and_tables():
-    # Drop existing tables to ensure the database schema matches the current models.
-    # WARNING: This will erase existing data — use only in development or when you
-    # intentionally want to reset the database.
-    SQLModel.metadata.drop_all(engine)
-    SQLModel.metadata.create_all(engine)
+def create_db_and_tables(drop_existing: bool = False):
+    """Crear la base de datos y todas las tablas."""
+    if drop_existing:
+        SQLModel.metadata.drop_all(engine)
+    first_time = not inspect(engine).has_table(Car.__tablename__)
+    if first_time:
+        logger.info("Creando todas las tablas en la base de datos.")
+        SQLModel.metadata.create_all(engine)
+
+
+# Función para inicializar la base de datos con datos de ejemplo
+def init_db():
+    """Inicializar la base de datos con datos de ejemplo."""
+    # Crear las tablas
+    drop_db = os.getenv("CLEAR_DB_ON_STARTUP", "False").lower() == "true"
+    create_db_and_tables(drop_existing=drop_db)
